AM335X: CM3: Fix issue in parameter parsing for command handlers
cmd_global_data.data which is passed to all the
command handlers (a8_lp_cmd3_handler etc) was always
pointing to the data in IPC registers irrespective
of the value of 'use_default_val'. Because of this,
the values populated to the members of ds0_data,
ds1_data, standby_data (Ex:wake_sources) are ignored
by the command handlers.

This issue was noticed while trying to get standby mode
working with only MPU_WAKE as wake source.

Signed-off-by: Satyanarayana Sandhya <sandhya.satyanarayana@ti.com>
Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>