AM335X: CM3: Ignore ON state settings for on-chip memories
ON state of the different on-chip memories, if required, can be
controlled from MPU itself. Moreover, so far we were just
restoring the pre-DS settings for these bits and hence there
is no functional difference in removing these from CM3. Doing
so also lets us repurpose some previous DS_PARAM bits. One of
the ON state bits in PD_PER is now repurposed for controlling
the second bank of OCMC memory in the next SoC.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>