CM3: Allow custom DS1 handler to change PER retention states.
Since the DS1 handler allows the user to change the PER power state,
it doesn't make sense to force all the retention states of that power
domain to be 0 (current behavior). Allow the passed in command to
change the retention states.

This is also to match up with the pd_*_stst_next functions so the code
in pm_handlers that duplicates it can be removed.

Signed-off-by: Russ Dill <Russ.Dill@ti.com>