CM3: Make LDO interface generic and separate am335x/am43xx
This makes the LDO interface generic and prepares for a
am335x/am43xx split. The removal of the ES1_0 check is
ok because the code that calls it already checks for that.

Future work might include a flag indicating that an LDO can
or cannot be put into retention.

Signed-off-by: Russ Dill <Russ.Dill@ti.com>