CM3: Enumerate command id.
Rather than code the magic value for the command id all over the place, create
an enum fer the command id and used the name of the command instead. This
makes the code more readable and easy to follow. Additionally, if proper
debug symbols are used, the command name can be viewed in the cmd_id
field when debugging with gdb.

Signed-off-by: Russ Dill <Russ.Dill@ti.com>