AM335X: CM3: Remove some non-essential module handling code
CM3 should be worried about only the essential modules.
The additional module handling code can and should be
removed.

Note that PSP kernels still require a EMIF enable because of a bug
in sleep33xx.S. This is moved to pm_handlers so that it can be
later conditionally executed only for those older kernels.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>
Signed-off-by: Russ Dill <Russ.Dill@ti.com>