CM3: AM437x: Add XTAL calculation code and support am437x
The AM437x CONFIG_STATUS register has two places where
the master xtal frequency can be indicated, and a bit showing
where software should read the value from.  Make a function
that works for both instead of open coding it in i2c.c. This
function works for both because the bit that indicates where
to read the value from is always 0 on am335x.

This was causing issues with a miscalculation of i2c fclk
frequency on AM437x.

Signed-off-by: Russ Dill <Russ.Dill@ti.com>